Apply authentication to agent apiserver endpoint
Antrea agent uses apiserver to interact with antctl. However, unlike the
controller, the API endpoint is exposed only locally which limits antctl
to local execution.
Additionally, we would like to reuse the listener for Prometheus metrics,
which would require external access to the API endpoint's metrics path.

Having authentication in a similar way to the controller's implementation
could help here as it will allow external exposure of the API endpoint.