Acquire xtables.lock before executing iptables-restore (#633)
We need to acquire xtables lock explicitly for iptables-restore to
prevent it from conflicting with iptables/iptables-restore which might
being called by kube-proxy. iptables supports "--wait" option and
go-iptables has enabled it. iptables-restore doesn't support the option
until 1.6.2, but it's not widely deployed yet.

Besides, this PR logs the error instead of the command when executing
iptables-restore fails.