Add default CPU resource requests for Antrea Pods in manifests
This is meant as an indication to the scheduler, to improve scheduling
of workload Pods.

It should not impact the likelihood of eviction of Antrea Pods. First,
CPU utilization does not seem to be used as an eviction signal. Second,
we already use a priorityClass of system-cluster-critical /
system-node-critical, which means that they are essentially treated as
Pods with a QoS class of Guaanteed (since K8s 1.14).

We do not define CPU limits as CPU usage can be very bursty (e.g. on
start up for antrea-agent, or if there is a burst of new flows for
antrea-ovs).

I don't think we should specify memory requests, as it is impossible to
come up with a number that would work for most use cases. For example,
memory footprint should be much smaller with for small clusters which no
NetworkPolicies.

Fixes #647