Fix return path for NodePort Service traffic in EKS
When Antrea is used to enforce NetworkPolicies in EKS, an additional
iptables rule is required in the mangle table, to ensure that return
NodePort Service traffic which was load-balanced to a Pod attached to a
secondary ENI takes the correct reverse path through eth0.

This requires communicating to the Antrea Agent that it is being run in
an EKS managed cluster, which we do through an environment variable
(ANTREA_CLOUD_EKS).

Some future improvements are possible: this rule is only necessary if
NodePort is enabled in AWS VPC CNI, and in some rare cases it may be
possible for the packet mark to differ from the default value (0x80).

This patch also improves the test-conformance-eks.sh script which was
used for testing. In particular it ensures that the script can be used
on macOS.

Fixes #678