Create and maintain connection store for Antrea flows in conntrack
Created connection store that stores the flows by polling conntrack module
every 5s. cxnStore updates the flow if it is already there or add a new
flow based on 5 tuple map. In addition, we add local pod information by
querying interfaceStore. We added ipCache in
interfaceStore for this purpose.

Unit tests and testing on local setup is done with custom apps.

Issue: #712