Fix antrea-agent not working with iptables>=1.6.2 (#698)
When antrea-agent ran with iptables>=1.6.2, it acquired the xtables lock
explicitly before calling iptables-restore. However, iptables-restore
will try to acquire the same lock and will fail immediately if it cannot
be acquired.

To make antrea-agent work with >=1.6.2 and <1.6.2 at the same time, this
PR detects the iptables version, uses "--wait" if it's greater than or
equal to 1.6.2, or acquire the lock explicitly.