Use Kustomize to generate release antrea-windows.yml
We follow the same model for antrea-windows.yml as we do for our other
YAML manifests. The YAML file can now be generated (using
hack/generate-manifest-windows.sh) in both "dev" and "release" modes,
using the appropriate container image for both.

antrea-windows.yml is added to the `make manifest` target and we verify
that it is up-to-date in CI.

The release antrea-window.yml will also automatically be uploaded as
part of release assets starting with Antrea v0.7.0.

The only notable changes in antrea-window.yml itself are that:
 - the ConfigMap used by the agent DaemonSet for Windows has been
   renamed to avoid conflicting with the ConfigMap used by the Linux
   agent DaemonSet
 - the missing enablePrometheus and apiPort conf parameters were added
   to antrea-agent.conf as I don't think they were omitted on purpose in
   the first place