Remove function callbacks for tracking Antrea agent metrics  (#781)
* Track Antrea agent metrics differently

Track Antrea agent metrics when the corresponding events happen rather
than through a callback. This is for optimizing scraping and also remove
dependencies on certain packages that are introduced by callbacks.
In addition, it fixes the dependency on metric types such as
NewCounterFunc, NewGuageFunc etc. and there by RawRegister, which is
being deprecated soon.

Fix #752

* Update prometheus.go