Add Github workflow for issue / PR lifecycle management
We use the stale action (https://github.com/actions/stale) to take care
of issue / PR state transitions and labelling.

The action is not quite as configurable as what we had in mind when we
wrote the issue management document, so we got rid of the
`lifecycle/rotten` label. But this label was only used to indicate that
the issue was being automatically closed.

Since we do not have a very high volume of new issues at this time, or a
very large backlog, I chose some more conservative time intervals for
transitions (issue becomes stale after 180 days, and is closed after an
extra 180 days of inactivity). These can be tuned in the future.