Add workaround for Agent readiness probe issue in Kind clusters
When deploying Antrea in Kind clusters, we sometimes observe that the
antrea-agent container does not become ready for a few minutes. This
seems to happen specifically when the antrea-agent starts and
establishes connections to the k8s apiserver or the Antrea Controller
before the br-phy bridge has been created. When this happens, the
creation of the br-phy bridge seems to break the existing connections,
and the container is eventually re-started after the TCP keepalive
timeout (~5 minutes with the Go HTTP library), which fixes the
issue. However, by this time, a few e2e tests have timed out and failed.

This workaround adds a small sleep (2 seconds) before starting the
antrea-agent process, which ensures that the antrea-ovs container has
enough time to create the br-phy bridge.

Fixes #801