Fix statefulset Pod recreate case
When deleting a StatefulSet Pod with 0 second grace period, the Pod
might be removed from the Kubernetes API very quickly and a new Pod is
created immediately, leading to kubelet processing the deletion of the
previous Pod and the add of the new Pod simultaneously, so CNI cannot
assume PodNamespace+PodName must be unique.

Previously antrea-agent computed the network device name by hashing
PodNamespace and PodName, the CNI DEL of the previous Pod could destroy
the network device of the new Pod if it's executed after the new one's
CNI ADD.

To fix it, antrea-agent must no longer use PodNamespace+PodName as the
identifier of the Pod's network device. This patch uses (sandbox)
container ID as part of the network device name, and uses container ID
as the key when caching interfaceConfig, therefore multiple Pods that
have the same name will be handled individually without interfering with
each other.

To be compatible with network devices created by previous versions, CNI
DEL handler doesn't assume how the network device name is computed and
always looks up network config by container ID.