Use Feature Gates to enable or disable features
There are a number of features that are being developed and likely to be
disabled by default in their early stage.

Instead of adding a temporary config for each feature and maintaining
them separately, this patch introduces Feature Gates to toggle the
features. It will be easier to choose code branch based on FeatureGates'
"Enabled" method and to promote features to beta and GA.