Change default tunnel type from VXLAN to Geneve
Geneve supports extensible encapsulation headers which allow Antrea to
add extra contexts to the encapsulated packets across Nodes. This is
required by the Antrea Traceflow feature (see issue #631) which is
expected to be released in Antrea 0.8.0.

When upgrading an existing K8s cluster that configured Antrea to use
VXLAN tunnel type, there are two options to handle the default tunnel
type change:
1) Edit the Antrea deployment YAML and explicitly set the
'tunnel_type" config option to 'vxlan` in 'antrea-agent.conf' of the
'antrea-config' ConfigMap. Then Antrea will keep using the VXLAN
tunnel type after upgrade.
2) Swtich to the new default tunnel type. Then after the new version
of Antrea agent starts, it will delete the old VXLAN tunnel port on the
OVS bridge and create a new tunnel port of the Geneve type. So, before
agents on all Nodes are upgraded and recreate the tunnel port, it is
possible some Nodes are using the VXLAN tunnel type, while some others
are using the Geneve tunnel type, and the traffic between the two sides
will be dropped.