Implements antctl query endpoint
Antctl query endpoint gives users the ability to filter network policies
relevant to a certain endpoint. The user is able to query internal
controller stores using the antrea controller apiserver through antctl.

Queries are answered by iterating over AddressGroups and
AppliedToGroups. While we considered adding Indexers to their respective
stores to speed-up queries, we decided that it was not sufficient to
justify the memory overhead observed in some cases.

Note, antctl query endpoint only works when run in the controller.
We are working on getting it to work out-of-cluster.

Fixes #974