Support log file size and number limits
klog does not enforce max log file size limit when the log file is not
specifiec by the --log_file argument. This commit sets klog.MaxSize to
the --log_file_max_size argument value when logging to file is enabled
and --log_file is not provided.
The commit also adds a new argument --log_file_max_num to define the
max number (per severity level) of log files to be kept. The file number
limit is enforced by Antrea code that periodically checks the log files
and deletes the oldest files to keep at most the specified max number of
log files.

Fixes: #788