Controller and Agent support for ClusterNetworkPolicy (#803)
Add ClusterNetworkPolicy event handlers to antrea-controller to handle ClusterNetworkPolicy CRD events.
Antrea-agent is updated to receive the events from antrea-controller and realize OF rules for corresponding CNP rules. A
priority module is added to antrea-agent to correctly allocate priorities to OF entries created based on the priority of the
CNP and the rule within. OpenFlow networkpolicy module is updated to handle drop rules and new OVS tables are created
in the pipeline to handle CNP rules.

Co-authored-by: Abhishek Raut <rauta@vmware.com>