Support for batch OVS flow updates
This is an optimization for NetworkPolicy OVS flows to be collectively installed
in agent restart case, so that those flows will rolled out at once and avoid potential
priority re-assignments.
In the future, mechanisms can also be added so that once this operation is done,
NetworkPolicy flows from the previous round are deleted immediately after.