[Windows] Restore lost host routes during agent initialization (#959)
When initialize the agent network on Windows node, the IP
configuration of uplink interface will be moved to OVS bridge
local interface. Which will cause the routes configured on
the uplink interface are lost.

This patch saves the routes configured on the uplink interface
before the IP configuration change and adds the saved routes on
OVS bridge local interface after that.

Signed-off-by: Rui Cao <rcao@vmware.com>