Fix golangci targets on macOS
Ever since goccy/go-graphviz was added as a dependency, running "make
golangci" (and related targets) has been broken on macOS. This is
because goccy/go-graphviz requires cgo to be enabled and go disables it
when cross-compiling (we set GOOS to linux in "make golangci" to ensure
that files with a linux build tag are verified by the linters).

We now run "make golangci" from a macOS worker in CI to avoid breaking
this again in the future.

In the future, we should also ensure that the linters are run on
Windows-specific files.

This commit also adds a few "missing" doc.go files.