Use Ubuntu 20.04 as base distribution in Antrea Docker image
The main reason for this update is picking up a more recent version of
glibc, as the one that ships with Ubuntu 18.04 can cause OVS to deadlock
(See #1022).

In this PR, we only update the distribution for the "main" Antrea Docker
image; other images, such as the ones we use for testing or for
deploying the Antrea Octant plugin, can be updated later if needed.

This is also a good opportunity to upgrade OVS daemons from 2.13.0 to
2.13.1, since the Docker build had to be updated anyway. For the sake of
simplicity, from now on we will only support building the base
openvswitch Docker image for OVS >= 2.13.0.

Fixes #1022