Increase watch timeout for Antrea apiserver
From the default 1800s (30 mins) to 2 hours.
I believe this may be a good idea for clusters with a large number of
Nodes and a large number of NetworkPolicies. K8s documentation also
recommends using a large value in production after verifying that
clients handle watch termination gracefully.

The verbosity level of the Controller log messages in case of watch
termination is also reduced to fix #1036. Unfortunetly there is no easy
way to make these log messages more useful (e.g. by correlating the
watcher object to the client).