Adding EKS node init script (#1057)
This is temporary solution to handle a race condition with cni
chaining mode. Basically if there are existing PODs before Antrea
is installled or in case of scale out if PODs are deployed before
Antrea daemonset is applied, those PODs will not be managed by
Antrea and hence network policies cannot be applied to those PODs.

With eks node init, pods are restarted once antrea config is detected.

Also updated the eks installation document and eks ci script. Removed
defaultMTU and serviceCIDR update from the eks ci script as its not needed
now. Default MTU will be auto discovered and with Antrea Proxy enabled for
networkPolicy mode, serviceCIDR is not required.

Co-authored-by: Rahul Jain <rahulj@rahulj-a01.vmware.com>