rotate generated self-signed certificate at half the expiration time
When rotating self-signed certificates generated by Antrea, we should
try to rotate them much earlier than 1 day before the expiration time.
This gives users time to manually intervene in the event that there are
issues with rotation. Also, shorter expiration times are always better,
especially if we are automatically rotating them ourselves.

This commit attempts to reduce the default rotation duration to half the
expiration time of the generated certificate, with a maximum rotate
duration of 1 year, in case the generated certificate expiration happens to be
really long.

Signed-off-by: Andrew Sy Kim <kim.andrewsy@gmail.com>