Remove unnecessary mutex in Antrea proxier code (#1138)
Reading the code, it seems that we never call syncProxyRules
directly. Instead we always invoke Run on the BoundedFrequencyRunner
object. But my understanding is that the BoundedFrequencyRunner will
serialize all the calls ("All runs will be async to the caller of
BoundedFrequencyRunner.Run, but multiple runs are serialized"). Given
that, and the fact that the data structures we are protecting are always
accessed through syncProxyRules, it seems that the mutex is not
necessary.

The change trackers (endpointsChangesTracker, servicesChangesTracker)
protect their own state using a mutex.