Traceflow: check that AntreaProxy is enabled for ClusterIP destinations
We already perform that check when a Service name is provided, but I
think we should do it as well for ClusterIP destinations. Otherwise, we
end up with an incorrect Traceflow result.

Logging this in the Agent is not ideal because you have to figure out
which logs to look at, but this is the only thing we can do right now.