Fix strongSwan configuration file with strongSwan >= version 5.7.0
Starting from version 5.7.0, strongSwan no longer supports specifying a
configuration parameter with the path delimited by dots in a
configuration file. This change breaks ovs-ipsec-monitor which writes
strongSwan configuration parameters with the dotted paths.
This commit applies an ovs-monitor-ipsec patch to fix the issue, as a
workaround before the patch is merged to OVS upstream.