Add "Has Changes" Github action (#1188)
This action can be used to determine whether a "diff" includes changes
to files whose paths are not part of a provided exclusion list.

A "diff" is defined as:
 * for a PR, the diff between the base SHA and the HEAD
 * for a push, the diff between the last SHA that was previously pushed
   to the branch and the HEAD

This action can be used to avoid running Kind tests for documentation
PRs. Unlike the "paths-ignore" option for Github workflows, we will
return a success status even when the tests are being skipped.