Add legacy networking API group for backward compatibility
Package networking is superseded by package controlplane in v0.9.3, to
keep backward compatibility with up to two minor versions of
antrea-agents, the API group should remain for two more minor releases.

This patch adds the APIs back and serves them with the same storage as
controlplane API. Also, kind tests are added to make sure basic
functionalities are not broken when only controller is upgraded.