Skip Docker builds in CI for documentation PRs
This extends https://github.com/vmware-tanzu/antrea/pull/1189 to skip
other CI jobs for documentation PRs. In particular, the Windows Docker
build, which takes more than 20 minutes to complete. Note that once the
documentation PR is merged, the 'push' event will trigger new Docker
images to be built and uploaded to dockerhub, which should be the
desired behavior.