Add network policy info to flow records
Network policy info (name and namespace), both ingress and egress
policies are added.

As part of this commit, we support exporting both flow records, for
the case of inter-Node flows. One from the source Node, where the
flow originates from and other from the destination Node of the flow,
where the destination Pod resides. The reason is that they contain
different NetworkPolicy info.