Download all artifacts from https://downloads.antrea.io
Fixes #1254

The SHA256 checksum verification for the OVS Windows binaries is no
longer required for security reasons, but it may be a good idea to keep
it (there is no real maintenance burden, we just need to update it
whenever the link is updated). It can help ensure that there is no issue
on our side and that the file that was uploaded to S3 by us is the
correct one. There is no burden on users. We can also remove that code
altogether.