Restrict read permissions for OVSDB
This patch removes "read" permissions for "others". The idea is to
improve security for OVSDB, which notably stores the PSK in plaintext
when using IPsec.

This is not perfect, as there is a time window between creation of the
db.conf file and the call to chmod.

See full discussion in #1292