NodePort, LoadBalancer and ClusterIP from k8s Node support for AntreaProxy on Linux
This PR implements:
 - The connection request of NodePort whose client is from remote or
localhost.
 - The connection request of LoadBalancer whose client is from remote
or localhost.
 - The connection request of ClusterIP whose client is from localhost.

For NodePort support, on each interface whose IP addresses can be
NodePort IP addresses, Linux TC is used to redirect the request packets
to Antrea gateway. For response packets, on interface Antrea gateway,
Linux TC is used to redirect the packets back to the interface where
the requests packets are from.

For LoadBalancer support, when client is from remote hosts, on default
route output interface, Linux TC is used to redirect the request
packets to Antrea gateway. For response packets, on interface Antrea
gateway,Linux TC is used to redirect the packets back to the default
route output interface. When client is from localhost, the request
packets are routed to Antrea gateway.

For ClusterIP support, the request packets are routed to Antrea gateway.

To support the Service traffic of above cases, there are main changes of
OVS pipeline.
- Add table serviceConntrackCommitTable 106 to perform SNAT for
Service traffic.
- Modify table hairpinSNATTable ID from 106 to 108.
- Modify table serviceHairpinTable ID from 29 to 23.
- Add table serviceConntrackTable 24 to transform SNATed connnections.
- Add table serviceClassifierTable 35 to classify the Service traffic.
- Add table serviceDstMacRewriteTable 75 to rewrite destination MAC
address of response Service packets.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>