Cache licensing info about dependencies in golicense workflow
The golicense workflow makes a lot of Github API calls and very often
runs into the limit for Github actions (1,000) which slows down the
workflow considerably and may even impact other workflows.

This enables caching for the workflow, using a hash of all the go.sum
files for the repo as the cache key, to avoid running the golicense tool
unless it is necessary. However, we always build the binary assets, as
this workflow is also meant to verify that
hack/release/prepare-assets.sh is not broken.

Fixes #1303