Ensure antrea-agent Pods not terminating in waitForAntreaDaemonSetPods
In some e2e tests that need to restart antrea-agent Pods, restarting
antrea-agent Pods is required and deleting Pods is used to trigger it.
However, DaemonSet.Status.NumberAvailable counts Pods even if they are
terminating. So waitForAntreaDaemonSetPods could succeed without waiting
for Pods being recreated if the process caught the termination signal
and delayed the termination.

This patch fixes it by making sure all antrea-agent Pods are not
terminating in waitForAntreaDaemonSetPods.