Check and ignore ANP and ACNP in agent when AntreaPolicy is not enabled
If the AntreaPolicy feature gate is not enabled on the Antrea Agent but
is enabled on the Antrea Controller (e.g. due to misconfiguration, or
Controller/Agents have not been all restarted after a configuration
change), Agent might crash with a segment fault when receiving and
processing ANP or ACNP messages from controller, so this commit adds
a check in Agent NetworkPolicyController to ingnore ANP and ACNP
messages when AntreaPolicy is not enabled.