Hardcode Github action actions/stale version to v3.0.8
Because of this issue: https://github.com/actions/stale/issues/173,
which is preventing the removal of the stale label when the issue
receives a new comment.

This commit also increases operations-per-run to 200, since we are
starting to have a decent amount of open issues and it seems we keep
running into the current limit (30).