Set tun_metadata0 for non-Traceflow traffics
This patch is to fix "wrong tun_metadata" issue, which affect Antrea
Traceflow feature.
Currently, OVS can not generates a megaflow that matches a field that
is not present, which caused matching wrong tun_metadata if the traffic
contains data with and without tun)metadata, and it is not sure if it
makes sense to support that (match a field that is not present). Also,
the current OpenFlow pipeline seems to assume that if tun_metadata0 is
not present, then the value of it is zero. But actually it is undefined
if the tun_metadata is not present.