Set tun_metadata0 for non-Traceflow traffics
This patch is to fix "wrong tun_metadata" issue, which affects Antrea
Traceflow feature.
Currently, OVS cannot generates a megaflow which matches a field that
is not present, which caused matching wrong tun_metadata if the traffic
contains data with and without tun_metadata, and it is not clear if it
makes sense to support that (match a field that is not present). Also,
the current OpenFlow pipeline seems to assume that if tun_metadata0 is
not present, then its value is zero. But actually the value is undefined
if the tun_metadata is not present.