Implement agent simulator as an independent bin
This patch implements agent simulator as add an independent bin, it runs with kubemark
which simulates kubelet, and mainly watches NetworkPolicies, AddressGroups and
AppliedToGroups from antrea controller and prints the events of these resources to log.
With agent simulator, we do not need to lauch large cluster for scale test.

The agent simulator uses labels and nodeaffinity to disable antrea-agent/antrea-controller
running on the simulated nodes, and we can add some taints to not allow other pods to run on
simulated nodes.

To use agent simulator, please refer to docs/antrea-agent-simulator.md.