Support Pod traffic shaping
The bandwidth plugin offered by the CNI plugin team can be used to
support ingress and egress traffic shaping for Kubernetes Pod, and
kubelet reads the bandwidth annotations on Pods and passes them as
parameters of CNI requests. To support traffic shaping in Antrea, the
bandwidth plugin should be appended to the chaining list of antrea cni
conf and its binary should be installed.

This patch does the above and adds an e2e test to ensure the integration
work as expected.