Introducing NodePortLocal in Antrea Agent
This commit introduces NodePortLocal (NPL) in Antrea agent.
With NPL, a pod pod port can be directly reached from external network through a port of the node on which the pod is running.
NPL programs IPTABLE rules to send incoming traffic on a specific port of a node to a POD port.
This information is exposed through annotation of the POD for consumption of other agents. An example of the pod annotation is given bellow.
metadata:
  annotations:
    npl.k8s.io/endpoints: '[{"podport":"8080","nodeip":"10.102.47.229","nodeport":"40002"}]'
More details about Node-Port-Local can be found in the design document: https://docs.google.com/document/d/1SoAKBvbbt2QMnGmt53E8ja0ZxCv_T4r1zLVnMzpykTg
To use this feature, following changes are required in Antrea configuration:
- Set NodePortLocal: true under featureGates in antrea-agent.conf in Antrea Configmap.
- In the Antrea Agent DaemonSet configure the port range for NPL (NPL_PORT_RANGE) in the env.  e.g.
        - name: NPL_PORT_RANGE
          value: 40000-46000
- Add capability for pod update in antrea-agent cluster role
Upcoming changes in future commits:
- Label based service filtering to select backend pods for NPL
- Option to select port range for NPL though Configmap
- Bootup sync : Compare programmed IPtable rules with pod annotation during bootup and add/delete new rules in iptable only if required.
    - Currently we are deleting all the rules and adding new rules for all the pods while booting up Antrea agent.

Signed-off-by: Hemant Shaw <hemant.shaw@avinetworks.com>
Signed-off-by: Manu Dilip Shah <manu.shah@avinetworks.com>
Signed-off-by: Monotosh Das <monotosh.das@avinetworks.com>
Signed-off-by: Shubham Chauhan <shubham@avinetworks.com>
Signed-off-by: Sudipta Biswas <sudipto@avinetworks.com>