Make controlplane NetworkPolicy cluster scoped (#1445)
The namespace and name of the controlplane NetworkPolicy were copied
from the namespace and name of its original NetworkPolicy. If a Antrea
NetworkPolicy and a K8s NetworkPolicy have the same namespace and name,
they would override each other when converting to controlplane
NetworkPolicy. Besides, similar to appliedToGroup and addressGroup, the
controlplane NetworkPolicy should be cluster scoped.

This patch makes the controlplane NetworkPolicy cluster scoped and
changes to use the UID of the original NetworkPolicy as its name to
ensure uniqueness.