Add "antctl proxy" support (#1452)
With this command antctl can operate as a reverse proxy for Antrea APIs,
similarly to "kubectl proxy" for the K8s APIs (we rely on the
k8s.io/kubectl Go module for the implementation). Thanks to this,
troubleshooting the APIs can become much easier and we hide complexity
from clients. One example is that it becomes much easier to use "go tool
pprof". The drawback here is that the HTTPS connection between the proxy
and the Antrea Agent / Controller is not secure (this can be fixed for
the Controller at least); in this it is very similar to the
supportbundle command implementation.

Fixes #1434