API change: controlplane v1beta2 (#1467)
* Upgrade ExternalEntity to v1alpha2 and controlplane API to v1beta2 (#1405)

* Upgrade ExternalEntity to v1alpha2 and update GroupMember struct

* Add more UT

* Add backward-comp UT testcases and address comments

* Address more comments

* Update controlplane.v1alpha1 to controlplane.v1alpha2

* Remove External Entity v1alpha1 support

* Fix yaml errors for v1alpha1 ExternalEntities

* Add upgrade tests into actions for controlplane v1beta2 branch

* Deprecate GroupMemberPod (#1416)

- Remove GroupMemberPod entirely
- Update AddressGroup and AppliedToGroup conversions
- Always add PodRef for GroupMember

* Make controlplane NetworkPolicy cluster scoped (#1445)

The namespace and name of the controlplane NetworkPolicy were copied
from the namespace and name of its original NetworkPolicy. If a Antrea
NetworkPolicy and a K8s NetworkPolicy have the same namespace and name,
they would override each other when converting to controlplane
NetworkPolicy. Besides, similar to appliedToGroup and addressGroup, the
controlplane NetworkPolicy should be cluster scoped.

This patch makes the controlplane NetworkPolicy cluster scoped and
changes to use the UID of the original NetworkPolicy as its name to
ensure uniqueness.

Co-authored-by: Quan Tian <qtian@vmware.com>